requirements.txt
asgiref==3.9.1
certifi==2025.8.3
Django==5.2.5
django-ckeditor-5==0.2.18
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
sib-api-v3-sdk==7.6.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
mysqlclient